requirements.txt
langchain-core==0.1.13
langchain==0.0.267
langchain-google-genai==0.0.6
langchain-community==0.0.10
streamlit==1.29.0
python-dotenv==1.0.0
google-generativeai==0.3.2
duckduckgo-search==3.9.9
pydantic==1.10.8
tenacity==8.2.3
regex==2023.10.3
watchdog==3.0.0